+++ src/lib.rs
@@ -6,4 +6,4 @@ pub mod json_utils;
 
 pub mod file_structs;
 pub mod url_structs;
-pub mod de;
+pub mod number;
