+++ src/url_structs.rs
@@ -1,31 +1,67 @@
 // Generated with schema_generator 0.0.0
-use crate::de::{Number, deserialize_optional_number};
+use crate::number::{Number, deserialize_optional_number};
 use serde::{Serialize, Deserialize};
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct TsTreeStandData {
+pub struct TstTreeStrata {
 	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
 	pub text: Option<String>,
-	#[serde(rename = "TreeStandDataDate", skip_serializing_if = "Option::is_none")]
-	pub ts_tree_stand_data_date: Option<Vec<TsTreeStandDataDate>>,
+	#[serde(rename = "TreeStratum", skip_serializing_if = "Option::is_none")]
+	pub tst_tree_stratum: Option<Vec<TstTreeStratum>>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct GdtPolygonGeometry {
+pub struct OpCutting {
 	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
 	pub text: Option<String>,
-	#[serde(rename = "pointProperty", skip_serializing_if = "Option::is_none")]
-	pub gml_point_property: Option<GmlpointProperty>,
-	#[serde(rename = "polygonProperty", skip_serializing_if = "Option::is_none")]
-	pub gml_polygon_property: Option<GmlpolygonProperty>,
+	#[serde(rename = "CuttingVolume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub op_cutting_volume: Option<Number>,
+	#[serde(rename = "Assortments", skip_serializing_if = "Option::is_none")]
+	pub op_assortments: Option<OpAssortments>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct GmlLinearRing {
+pub struct Gmlexterior {
 	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
 	pub text: Option<String>,
-	#[serde(rename = "coordinates", skip_serializing_if = "Option::is_none")]
-	pub gml_coordinates: Option<String>,
+	#[serde(rename = "LinearRing", skip_serializing_if = "Option::is_none")]
+	pub gml_linear_ring: Option<GmlLinearRing>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct GmlpolygonProperty {
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "Polygon", skip_serializing_if = "Option::is_none")]
+	pub gml_polygon: Option<GmlPolygon>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct Gmlinterior {
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "LinearRing", skip_serializing_if = "Option::is_none")]
+	pub gml_linear_ring: Option<GmlLinearRing>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct GmlpointProperty {
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "Point", skip_serializing_if = "Option::is_none")]
+	pub gml_point: Option<GmlPoint>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct GmlPolygon {
+	#[serde(rename = "@srsName", skip_serializing_if = "Option::is_none")]
+	pub srs_name: Option<String>,
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "exterior", skip_serializing_if = "Option::is_none")]
+	pub gml_exterior: Option<Gmlexterior>,
+	#[serde(rename = "interior", skip_serializing_if = "Option::is_none")]
+	pub gml_interior: Option<Gmlinterior>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -39,15 +75,13 @@ pub struct OpProposalData {
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct GmlPolygon {
-	#[serde(rename = "@srsName", skip_serializing_if = "Option::is_none")]
-	pub srs_name: Option<String>,
+pub struct StIdentifier {
 	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
 	pub text: Option<String>,
-	#[serde(rename = "exterior", skip_serializing_if = "Option::is_none")]
-	pub gml_exterior: Option<Gmlexterior>,
-	#[serde(rename = "interior", skip_serializing_if = "Option::is_none")]
-	pub gml_interior: Option<Gmlinterior>,
+	#[serde(rename = "IdentifierType", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub co_identifier_type: Option<Number>,
+	#[serde(rename = "IdentifierValue", skip_serializing_if = "Option::is_none")]
+	pub co_identifier_value: Option<String>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -91,6 +125,20 @@ pub struct ForestPropertyData {
 }
 
 #[derive(Serialize, Deserialize, Debug)]
+pub struct StSpecialFeature {
+	#[serde(rename = "@id")]
+	pub id: String,
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "MainFeature", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub sf_main_feature: Option<Number>,
+	#[serde(rename = "FeatureCode", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub sf_feature_code: Option<Number>,
+	#[serde(rename = "FeatureAdditionalCode", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub sf_feature_additional_code: Option<Number>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
 pub struct OpAssortment {
 	#[serde(rename = "@id")]
 	pub id: String,
@@ -105,97 +153,27 @@ pub struct OpAssortment {
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct Gmlinterior {
+pub struct OpCompletionData {
 	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
 	pub text: Option<String>,
-	#[serde(rename = "LinearRing", skip_serializing_if = "Option::is_none")]
-	pub gml_linear_ring: Option<GmlLinearRing>,
+	#[serde(rename = "CompletionDate", skip_serializing_if = "Option::is_none")]
+	pub op_completion_date: Option<String>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct StIdentifier {
+pub struct OpOperations {
 	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
 	pub text: Option<String>,
-	#[serde(rename = "IdentifierType", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub co_identifier_type: Option<Number>,
-	#[serde(rename = "IdentifierValue", skip_serializing_if = "Option::is_none")]
-	pub co_identifier_value: Option<String>,
+	#[serde(rename = "Operation", skip_serializing_if = "Option::is_none")]
+	pub op_operation: Option<Vec<OpOperation>>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct StSpecialFeatures {
+pub struct GmlLinearRing {
 	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
 	pub text: Option<String>,
-	#[serde(rename = "SpecialFeature", skip_serializing_if = "Option::is_none")]
-	pub st_special_feature: Option<Vec<StSpecialFeature>>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OpOperation {
-	#[serde(rename = "@id")]
-	pub id: String,
-	#[serde(rename = "@mainType")]
-	pub main_type: String,
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "OperationType", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub op_operation_type: Option<Number>,
-	#[serde(rename = "ProposalData", skip_serializing_if = "Option::is_none")]
-	pub op_proposal_data: Option<OpProposalData>,
-	#[serde(rename = "Cutting", skip_serializing_if = "Option::is_none")]
-	pub op_cutting: Option<OpCutting>,
-	#[serde(rename = "CompletionData", skip_serializing_if = "Option::is_none")]
-	pub op_completion_data: Option<OpCompletionData>,
-	#[serde(rename = "DataSource", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub co_data_source: Option<Number>,
-	#[serde(rename = "Silviculture", skip_serializing_if = "Option::is_none")]
-	pub op_silviculture: Option<OpSilviculture>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct TstTreeStrata {
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "TreeStratum", skip_serializing_if = "Option::is_none")]
-	pub tst_tree_stratum: Option<Vec<TstTreeStratum>>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct TssTreeStandSummary {
-	#[serde(rename = "@id")]
-	pub id: String,
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "MeanAge", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_mean_age: Option<Number>,
-	#[serde(rename = "BasalArea", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_basal_area: Option<Number>,
-	#[serde(rename = "StemCount", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_stem_count: Option<Number>,
-	#[serde(rename = "MeanDiameter", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_mean_diameter: Option<Number>,
-	#[serde(rename = "MeanHeight", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_mean_height: Option<Number>,
-	#[serde(rename = "Volume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_volume: Option<Number>,
-	#[serde(rename = "SawLogVolume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_saw_log_volume: Option<Number>,
-	#[serde(rename = "PulpWoodVolume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_pulp_wood_volume: Option<Number>,
-	#[serde(rename = "VolumeGrowth", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_volume_growth: Option<Number>,
-	#[serde(rename = "LeafBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_leaf_biomass: Option<Number>,
-	#[serde(rename = "BranchBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_branch_biomass: Option<Number>,
-	#[serde(rename = "StemBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_stem_biomass: Option<Number>,
-	#[serde(rename = "StumpBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_stump_biomass: Option<Number>,
-	#[serde(rename = "DevelopmentClass", skip_serializing_if = "Option::is_none")]
-	pub tss_development_class: Option<String>,
-	#[serde(rename = "MainTreeSpecies", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tss_main_tree_species: Option<Number>,
+	#[serde(rename = "coordinates", skip_serializing_if = "Option::is_none")]
+	pub gml_coordinates: Option<String>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -215,159 +193,11 @@ pub struct DtsDeadTreeStratum {
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct StSpecialFeature {
-	#[serde(rename = "@id")]
-	pub id: String,
+pub struct StSpecialFeatures {
 	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
 	pub text: Option<String>,
-	#[serde(rename = "MainFeature", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub sf_main_feature: Option<Number>,
-	#[serde(rename = "FeatureCode", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub sf_feature_code: Option<Number>,
-	#[serde(rename = "FeatureAdditionalCode", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub sf_feature_additional_code: Option<Number>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct StIdentifiers {
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "Identifier", skip_serializing_if = "Option::is_none")]
-	pub st_identifier: Option<Vec<StIdentifier>>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct StStands {
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "Stand", skip_serializing_if = "Option::is_none")]
-	pub st_stand: Option<Vec<StStand>>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct TstTreeStratum {
-	#[serde(rename = "@id")]
-	pub id: String,
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "StratumNumber", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_stratum_number: Option<Number>,
-	#[serde(rename = "TreeSpecies", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_tree_species: Option<Number>,
-	#[serde(rename = "Storey", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_storey: Option<Number>,
-	#[serde(rename = "Age", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_age: Option<Number>,
-	#[serde(rename = "BasalArea", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_basal_area: Option<Number>,
-	#[serde(rename = "MeanDiameter", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_mean_diameter: Option<Number>,
-	#[serde(rename = "MeanHeight", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_mean_height: Option<Number>,
-	#[serde(rename = "Volume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_volume: Option<Number>,
-	#[serde(rename = "StemCount", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_stem_count: Option<Number>,
-	#[serde(rename = "SawLogVolume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_saw_log_volume: Option<Number>,
-	#[serde(rename = "PulpWoodVolume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_pulp_wood_volume: Option<Number>,
-	#[serde(rename = "VolumeGrowth", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_volume_growth: Option<Number>,
-	#[serde(rename = "LeafBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_leaf_biomass: Option<Number>,
-	#[serde(rename = "BranchBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_branch_biomass: Option<Number>,
-	#[serde(rename = "StemBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_stem_biomass: Option<Number>,
-	#[serde(rename = "StumpBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub tst_stump_biomass: Option<Number>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct GmlpointProperty {
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "Point", skip_serializing_if = "Option::is_none")]
-	pub gml_point: Option<GmlPoint>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct GmlpolygonProperty {
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "Polygon", skip_serializing_if = "Option::is_none")]
-	pub gml_polygon: Option<GmlPolygon>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct TsTreeStandDataDate {
-	#[serde(rename = "@type")]
-	pub ts_tree_stand_data_date_type: String,
-	#[serde(rename = "@date")]
-	pub date: String,
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "TreeStrata", skip_serializing_if = "Option::is_none")]
-	pub tst_tree_strata: Option<TstTreeStrata>,
-	#[serde(rename = "TreeStandSummary", skip_serializing_if = "Option::is_none")]
-	pub tss_tree_stand_summary: Option<TssTreeStandSummary>,
-	#[serde(rename = "DeadTreeStrata", skip_serializing_if = "Option::is_none")]
-	pub dts_dead_tree_strata: Option<DtsDeadTreeStrata>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Gmlexterior {
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "LinearRing", skip_serializing_if = "Option::is_none")]
-	pub gml_linear_ring: Option<GmlLinearRing>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OpSilviculture {
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct StStand {
-	#[serde(rename = "@id")]
-	pub id: String,
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "StandBasicData", skip_serializing_if = "Option::is_none")]
-	pub st_stand_basic_data: Option<StStandBasicData>,
-	#[serde(rename = "TreeStandData", skip_serializing_if = "Option::is_none")]
-	pub ts_tree_stand_data: Option<TsTreeStandData>,
-	#[serde(rename = "Operations", skip_serializing_if = "Option::is_none")]
-	pub op_operations: Option<OpOperations>,
-	#[serde(rename = "SpecialFeatures", skip_serializing_if = "Option::is_none")]
-	pub st_special_features: Option<StSpecialFeatures>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OpCutting {
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "CuttingVolume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
-	pub op_cutting_volume: Option<Number>,
-	#[serde(rename = "Assortments", skip_serializing_if = "Option::is_none")]
-	pub op_assortments: Option<OpAssortments>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct DtsDeadTreeStrata {
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "DeadTreeStratum", skip_serializing_if = "Option::is_none")]
-	pub dts_dead_tree_stratum: Option<Vec<DtsDeadTreeStratum>>,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OpCompletionData {
-	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
-	pub text: Option<String>,
-	#[serde(rename = "CompletionDate", skip_serializing_if = "Option::is_none")]
-	pub op_completion_date: Option<String>,
+	#[serde(rename = "SpecialFeature", skip_serializing_if = "Option::is_none")]
+	pub st_special_feature: Option<Vec<StSpecialFeature>>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -419,6 +249,126 @@ pub struct StStandBasicData {
 }
 
 #[derive(Serialize, Deserialize, Debug)]
+pub struct DtsDeadTreeStrata {
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "DeadTreeStratum", skip_serializing_if = "Option::is_none")]
+	pub dts_dead_tree_stratum: Option<Vec<DtsDeadTreeStratum>>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct StIdentifiers {
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "Identifier", skip_serializing_if = "Option::is_none")]
+	pub st_identifier: Option<Vec<StIdentifier>>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct TsTreeStandDataDate {
+	#[serde(rename = "@type")]
+	pub ts_tree_stand_data_date_type: String,
+	#[serde(rename = "@date")]
+	pub date: String,
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "TreeStrata", skip_serializing_if = "Option::is_none")]
+	pub tst_tree_strata: Option<TstTreeStrata>,
+	#[serde(rename = "TreeStandSummary", skip_serializing_if = "Option::is_none")]
+	pub tss_tree_stand_summary: Option<TssTreeStandSummary>,
+	#[serde(rename = "DeadTreeStrata", skip_serializing_if = "Option::is_none")]
+	pub dts_dead_tree_strata: Option<DtsDeadTreeStrata>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct OpOperation {
+	#[serde(rename = "@id")]
+	pub id: String,
+	#[serde(rename = "@mainType")]
+	pub main_type: String,
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "OperationType", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub op_operation_type: Option<Number>,
+	#[serde(rename = "ProposalData", skip_serializing_if = "Option::is_none")]
+	pub op_proposal_data: Option<OpProposalData>,
+	#[serde(rename = "Cutting", skip_serializing_if = "Option::is_none")]
+	pub op_cutting: Option<OpCutting>,
+	#[serde(rename = "CompletionData", skip_serializing_if = "Option::is_none")]
+	pub op_completion_data: Option<OpCompletionData>,
+	#[serde(rename = "DataSource", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub co_data_source: Option<Number>,
+	#[serde(rename = "Silviculture", skip_serializing_if = "Option::is_none")]
+	pub op_silviculture: Option<OpSilviculture>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct TsTreeStandData {
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "TreeStandDataDate", skip_serializing_if = "Option::is_none")]
+	pub ts_tree_stand_data_date: Option<Vec<TsTreeStandDataDate>>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct TssTreeStandSummary {
+	#[serde(rename = "@id")]
+	pub id: String,
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "MeanAge", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_mean_age: Option<Number>,
+	#[serde(rename = "BasalArea", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_basal_area: Option<Number>,
+	#[serde(rename = "StemCount", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_stem_count: Option<Number>,
+	#[serde(rename = "MeanDiameter", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_mean_diameter: Option<Number>,
+	#[serde(rename = "MeanHeight", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_mean_height: Option<Number>,
+	#[serde(rename = "Volume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_volume: Option<Number>,
+	#[serde(rename = "SawLogVolume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_saw_log_volume: Option<Number>,
+	#[serde(rename = "PulpWoodVolume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_pulp_wood_volume: Option<Number>,
+	#[serde(rename = "VolumeGrowth", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_volume_growth: Option<Number>,
+	#[serde(rename = "LeafBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_leaf_biomass: Option<Number>,
+	#[serde(rename = "BranchBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_branch_biomass: Option<Number>,
+	#[serde(rename = "StemBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_stem_biomass: Option<Number>,
+	#[serde(rename = "StumpBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_stump_biomass: Option<Number>,
+	#[serde(rename = "DevelopmentClass", skip_serializing_if = "Option::is_none")]
+	pub tss_development_class: Option<String>,
+	#[serde(rename = "MainTreeSpecies", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tss_main_tree_species: Option<Number>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct OpSilviculture {
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct StStand {
+	#[serde(rename = "@id")]
+	pub id: String,
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "StandBasicData", skip_serializing_if = "Option::is_none")]
+	pub st_stand_basic_data: Option<StStandBasicData>,
+	#[serde(rename = "TreeStandData", skip_serializing_if = "Option::is_none")]
+	pub ts_tree_stand_data: Option<TsTreeStandData>,
+	#[serde(rename = "Operations", skip_serializing_if = "Option::is_none")]
+	pub op_operations: Option<OpOperations>,
+	#[serde(rename = "SpecialFeatures", skip_serializing_if = "Option::is_none")]
+	pub st_special_features: Option<StSpecialFeatures>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
 pub struct GmlPoint {
 	#[serde(rename = "@srsName", skip_serializing_if = "Option::is_none")]
 	pub srs_name: Option<String>,
@@ -429,11 +379,51 @@ pub struct GmlPoint {
 }
 
 #[derive(Serialize, Deserialize, Debug)]
-pub struct OpOperations {
+pub struct StStands {
 	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
 	pub text: Option<String>,
-	#[serde(rename = "Operation", skip_serializing_if = "Option::is_none")]
-	pub op_operation: Option<Vec<OpOperation>>,
+	#[serde(rename = "Stand", skip_serializing_if = "Option::is_none")]
+	pub st_stand: Option<Vec<StStand>>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct TstTreeStratum {
+	#[serde(rename = "@id")]
+	pub id: String,
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "StratumNumber", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_stratum_number: Option<Number>,
+	#[serde(rename = "TreeSpecies", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_tree_species: Option<Number>,
+	#[serde(rename = "Storey", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_storey: Option<Number>,
+	#[serde(rename = "Age", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_age: Option<Number>,
+	#[serde(rename = "BasalArea", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_basal_area: Option<Number>,
+	#[serde(rename = "MeanDiameter", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_mean_diameter: Option<Number>,
+	#[serde(rename = "MeanHeight", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_mean_height: Option<Number>,
+	#[serde(rename = "Volume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_volume: Option<Number>,
+	#[serde(rename = "StemCount", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_stem_count: Option<Number>,
+	#[serde(rename = "SawLogVolume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_saw_log_volume: Option<Number>,
+	#[serde(rename = "PulpWoodVolume", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_pulp_wood_volume: Option<Number>,
+	#[serde(rename = "VolumeGrowth", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_volume_growth: Option<Number>,
+	#[serde(rename = "LeafBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_leaf_biomass: Option<Number>,
+	#[serde(rename = "BranchBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_branch_biomass: Option<Number>,
+	#[serde(rename = "StemBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_stem_biomass: Option<Number>,
+	#[serde(rename = "StumpBiomass", deserialize_with = "deserialize_optional_number", skip_serializing_if = "Option::is_none", default)]
+	pub tst_stump_biomass: Option<Number>,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
@@ -442,5 +432,15 @@ pub struct OpAssortments {
 	pub text: Option<String>,
 	#[serde(rename = "Assortment", skip_serializing_if = "Option::is_none")]
 	pub op_assortment: Option<Vec<OpAssortment>>,
+}
+
+#[derive(Serialize, Deserialize, Debug)]
+pub struct GdtPolygonGeometry {
+	#[serde(rename = "$text", skip_serializing_if = "Option::is_none")]
+	pub text: Option<String>,
+	#[serde(rename = "pointProperty", skip_serializing_if = "Option::is_none")]
+	pub gml_point_property: Option<GmlpointProperty>,
+	#[serde(rename = "polygonProperty", skip_serializing_if = "Option::is_none")]
+	pub gml_polygon_property: Option<GmlpolygonProperty>,
 }
